Update interdependent packages with breaking changes.
This advances a few versions of `System.IO.Abstractions`, which keeps making breaking changes to the interfaces, and also a few unstable versions of Mutagen which depends on `System.IO.Abstractions` and therefore must also update.

Used the oldest possible version of Mutagen that is compatible with the new version of IO, to minimize the number of simultaneous changes.