Switch `GameSetup` to use game-agnostic method of reading masters.
Mutagen has a [convenience method](https://mutagen-modding.github.io/Mutagen/Big-Cheat-Sheet/#get-list-of-masters-from-a-mod) to perform this exact task more precisely, but we also want to really try to load the mod in order to do basic validation.

Unfortunately this does not have any async support, but for the moment it seems to be fast enough not to seriously delay startup.

This probably fixes the issue described in #202, although there may be many other FO4-related issues to follow. It definitely does fix the `ObjectDisposedException` encountered in Mutagen 0.45 when calling `CreateFromBinaryOverlay`.