Start of evaluation model.
Updates/cleans up some existing data types and adds new ones to support the idea of an evaluation or "dry run" to spell out what will happen to an NPC before any of the work is done.

Assets (facegen files) are in the data model but not implemented in the explainer yet.