Add preliminary mod indexing/analysis.
Currently only works for Mod Organizer, and is a bit naive about mod names, but does a decent job of generating unique keys and grouping together components of the same mod.

To improve this, could once again look at MO2 download metadata - knowing this isn't reliable for WJ installations - or try to use the Nexus API and contend with latency and rate limits (maybe OK if it's done in the background).