Implement mod asset mapping.
This is now a class and builder method that can generically operate off any mod registry and any asset list. It is _not_ intended to completely replace file/archiving indexing, only to provide highly definitive results on the mods providing each plugin, so as to reliably search for facegen assets in the providing mod and give immediate in-app feedback when none exists.

Includes some dependency refactoring to make the Mod Manager abstraction and creation a little more generic, although we still do not have environment detection in place and are using a hack for the instance directory.